+++ app/src/main/java/com/blockeq/stellarwallet/fragments/tabs/MyOffersTabFragment.kt
@@ -73,7 +73,7 @@ class MyOffersTabFragment : Fragment(), OnDeleteRequest, SwipeRefreshLayout.OnRe
                     val currencyBuy = Currency(1, buyingCode, "$buyingCode COIN", 0.0, null)
                     val sellingCode : String = AssetUtil.getCode(it.selling)!!
                     val currencySelling = Currency(2, sellingCode, "$sellingCode COIN", 0.0, null)
-                    myOffers.add(MyOffer(it.id.toInt(), Date(), currencySelling, currencyBuy, it.amount.toFloat(), it.price.toFloat()))
+                    myOffers.add(MyOffer(it.id.toInt(), Date(), currencySelling, currencyBuy, it.amount.toFloat(), it.amount.toFloat() * it.price.toFloat()))
                     id++
                 }
                 myOffersAdapter.notifyDataSetChanged()