+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.5] 2018-12-19
+## Fixed
+- GH-171 Toast crash from background thread
+- GH-173 Fixes `NumberFormatException` with empty buying price
+
+## [2.0.4] 2018-12-17
+## Fixed
+- Formatted with 4 decimals the amounts under myOffers list
+- Added -- for walletAdapter that are an inflation set
+- Used the availableBalance in the TradeTabFragment when the selling coin is LMX
+- Set a maximum of 7 decimals amount in the SendActivity and AddressActivity
+
+## [Internal-2.0.3] 2018-12-14
+## Fixed
+- Fixed the submit order button which was disabled during 100% trade.
+- [NumberFormatException] entering a invalid number (,) to the editText in the tradeView
+
 ## [Internal-2.0.2] 2018-12-14
 ## Fixed
 - Re-fix: Market trade amount in second text field is not a dark enough font color